requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
transformers==4.41.2
torch==2.3.0
pytz==2024.1
lxml_html_clean==0.1.1
Flask==3.0.3
gunicorn==22.0.0